requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pandas
matplotlib
openpyxl
pydantic
